requirements.txt
asgiref==3.6.0
Django==4.0.10
psycopg2-binary==2.9.3
sqlparse==0.4.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.54.0
environs[django]==9.5.0
pandas==2.0.3
geopy==2.4.1
spacy==3.7.2
django-htmx==1.17.3
requests==2.31.0
scrapy==2.11.2
scrapyd==1.4.3
python-scrapyd-api==2.1.2
Pillow==10.4.0